Crate ordered_iter

Ordered iterators.

Structs

InnerJoinMapIterator

InnerJoinMapSetIterator

InnerJoinSetIterator

OuterJoinIterator

Traits

OrderedMapIterator
Allows an iterator to be do an inner join with another iterator to combine their values or filter based on their keys. this trait is applied to an iterator over a map like structure

OrderedSetIterator
Allows an iterator to be do an inner join with another iterator to combine their values or filter based on their keys. this trait is applied to an iterator over a set like structure