Trait ordered_iter::OrderedSetIterator

pub trait OrderedSetIterator: Iterator + Sized {
    fn inner_join_map<I>(self, map: I) -> InnerJoinMapSetIterator<I, Self> where I: OrderedMapIterator<Key=Self::Item> { ... }
    fn inner_join_set<I>(self, map: I) -> InnerJoinSetIterator<Self, I> where I: OrderedSetIterator<Item=Self::Item> { ... }
}

Allows an iterator to be do an inner join with another iterator to combine their values or filter based on their keys. this trait is applied to an iterator over a set like structure

Provided Methods

fn inner_join_map<I>(self, map: I) -> InnerJoinMapSetIterator<I, Self> where I: OrderedMapIterator<Key=Self::Item>

join two ordered maps together

fn inner_join_set<I>(self, map: I) -> InnerJoinSetIterator<Self, I> where I: OrderedSetIterator<Item=Self::Item>

filter an ordered map with an ordered set